Project restructuring
Changed the package main package from ke.co.slick to
ke.co.slick.researchapp. This is so that the package name can be
consistent with the applicationID and also the package name in
AndroidManifest.xml

Also added a dependency injection templates with an empty component and
code to instantiate the generated component class.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>