UI Changes
Changed the layout for the results for a PubAg search query.
First, I removed the "issn" and "source" textViews since they just
didn't look right. Also, their removal is so that the results for
PubAg can have a similar layout with those for Springer Nature.

Secondly, I reduced the margins between lines to zero just as I did with
the Springer Nature api.

Thirdly, the abstract text is now a singleLine instead of having a max
of 3 lines.

Lastly, changed the font sizes to distinguish between the different
elements.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>